#22 Merge api and spec submodules
