--- tests/test_init.py
@@ -1,26 +1,0 @@
-# dependencies
-import xarrayspecs as xs
-
-
-def test_namespace() -> None:
-    # submodules
-    xs.api
-    xs.convert
-    xs.spec
-    # aliases
-    xs.AsDataArray
-    xs.AsDataset
-    xs.AsDataTree
-    xs.Dims
-    xs.Dtype
-    xs.Type
-    xs.asdataarray
-    xs.asdataset
-    xs.asdatatree
-    xs.attrs
-    xs.dims
-    xs.dtype
-    xs.name
-    xs.node
-    xs.type
-    xs.use
