+++ xarrayspecs/api.py
@@ -5,20 +5,29 @@ __all__ = [
     "asdataarray",
     "asdataset",
     "asdatatree",
+    "attrs",
+    "dims",
+    "dtype",
+    "name",
+    "node",
+    "type",
+    "use",
 ]
 
 # standard library
-from collections.abc import Callable
-from typing import Any, Protocol, TypeVar, overload
+from collections.abc import Callable, Hashable, Iterable, Mapping
+from typing import Any, Literal, Protocol, TypeVar, overload
 
 # dependencies
 import xarray as xr
+from typespecs import Spec
 from typing_extensions import ParamSpec
 from .convert import to_dataarray, to_dataset, to_datatree, to_specframe
 
 # type hints
 P = ParamSpec("P")
 T = TypeVar("T", bound=xr.DataArray | xr.Dataset | xr.DataTree)
+Use = Literal["attr", "attrs", "coord", "coords", "data", "name", "type", "vars"]
 
 
 class HasType(Protocol[P, T]):
@@ -111,3 +120,59 @@ def asdatatree(obj: Other[P], /) -> xr.DataTree: ...
 def asdatatree(obj: Any, /) -> Any:
     """Convert given Xarray specifications to an Xarray DataTree."""
     return to_datatree(to_specframe(obj))
+
+
+@overload
+def attrs(attrs: Mapping[Any, Any] | None, /) -> Spec: ...
+@overload
+def attrs(**attrs: Any) -> Spec: ...
+
+
+def attrs(*args: Any, **kwargs: Any) -> Spec:
+    """Returns a type specification for Xarray attributes."""
+    if len(args) == 0:
+        return Spec(xarray_attrs=kwargs)
+
+    if len(args) == 1 and not kwargs:
+        return Spec(xarray_attrs=args[0])
+
+    raise ValueError("Cannot create type specification.")
+
+
+@overload
+def dims(dims: Iterable[Hashable] | None, /) -> Spec: ...
+@overload
+def dims(*dims: Hashable) -> Spec: ...
+
+
+def dims(*args: Any) -> Spec:
+    """Returns a type specification for Xarray dimensions."""
+    if len(args) == 1:
+        return Spec(xarray_dims=args[0])
+    else:
+        return Spec(xarray_dims=args)
+
+
+def dtype(dtype: Any | None, /) -> Spec:
+    """Returns a type specification for Xarray data type."""
+    return Spec(xarray_dtype=dtype)
+
+
+def name(name: Hashable | None, /) -> Spec:
+    """Returns a type specification for Xarray name."""
+    return Spec(xarray_name=name)
+
+
+def node(node: str | None, /) -> Spec:
+    """Returns a type specification for Xarray node."""
+    return Spec(xarray_node=node)
+
+
+def type(type: Callable[..., Any] | None, /) -> Spec:
+    """Returns a type specification for Xarray type."""
+    return Spec(xarray_type=type)
+
+
+def use(use: Use | None, /) -> Spec:
+    """Returns a type specification for Xarray use."""
+    return Spec(xarray_use=use)
